feat: refactor config to be internal, have WithAccessControlParams pass params individually without the struct
