+++ tests/listusers/listusers_test.go
@@ -19,8 +19,8 @@ import (
 	"google.golang.org/grpc"
 
 	"github.com/openfga/openfga/cmd/run"
+	"github.com/openfga/openfga/internal/server/config"
 	"github.com/openfga/openfga/pkg/logger"
-	"github.com/openfga/openfga/pkg/server/config"
 	"github.com/openfga/openfga/pkg/testutils"
 	"github.com/openfga/openfga/pkg/tuple"
 	"github.com/openfga/openfga/pkg/typesystem"
